refactor(blocks): consolidate stat card styles into stat-card-chart
Remove stat-card-stat-styles.ts, move shared types and Tailwind class names into stat-card-chart, and replace inline chart height styles with cn size presets (sm/md/lg). Sync block registry outputs.
