+++ apps/web/blocks/_shared/files/components/stat-card-choropleth-hover-bridge.tsx
@@ -3,7 +3,7 @@
 import { useChoropleth } from "@bklitui/ui/charts";
 import { useEffect } from "react";
 import { computeVisitorTrend, getVisitorValue } from "../data/visitors";
-import type { StatCardHoverState } from "./stat-card-stat-styles";
+import type { StatCardHoverState } from "./stat-card-chart";
 
 /** Syncs hovered choropleth feature into stat card NumberFlow and trend badge. */
 export function StatCardChoroplethHoverBridge({
