+++ apps/web/blocks/stat-card-line-01/files/components/stat-card-chart.tsx
@@ -3,28 +3,41 @@
 import type { ReactNode } from "react";
 import { cn } from "@/lib/utils";
 
+export interface StatCardHoverState {
+  value: number | null;
+  label: string | null;
+  trend: number | null;
+}
+
+export const statCardValueClassName =
+  "text-3xl font-semibold leading-none tracking-tight";
+
+export const statCardLabelClassName = "mt-0 text-xs";
+
+export const statCardChartHeights = {
+  sm: "[--stat-card-chart-h:96px]",
+  md: "[--stat-card-chart-h:190px]",
+  lg: "[--stat-card-chart-h:420px]",
+} as const;
+
 /** Bleeds charts edge-to-edge inside stat card content padding. */
 export function StatCardChart({
   children,
   className,
-  height = 96,
+  size = "sm",
 }: {
   children: ReactNode;
   className?: string;
-  height?: number;
+  size?: keyof typeof statCardChartHeights;
 }) {
   return (
     <div
       className={cn(
         "relative -mx-4 -mb-3 overflow-hidden",
         "[&_.relative.w-full]:aspect-auto! [&_.relative.w-full]:h-[var(--stat-card-chart-h)]!",
+        statCardChartHeights[size],
         className
       )}
-      style={
-        {
-          "--stat-card-chart-h": `${height}px`,
-        } as React.CSSProperties
-      }
     >
       {children}
     </div>
