+++ apps/web/blocks/stat-card-line-01/files/components/stat-card-hover-bridge.tsx
@@ -2,9 +2,9 @@
 
 import { useChart } from "@bklitui/ui/charts";
 import { useEffect } from "react";
-import type { StatCardHoverState } from "./stat-card-stat-styles";
+import type { StatCardHoverState } from "./stat-card-chart";
 
-export type { StatCardHoverState } from "./stat-card-stat-styles";
+export type { StatCardHoverState } from "./stat-card-chart";
 
 export function formatStatCardMonth(date: Date) {
   return date.toLocaleDateString("en-US", { month: "short" });
