--- apps/web/blocks/stat-card-line-01/files/components/stat-card-stat-styles.ts
@@ -1,10 +1,0 @@
-export interface StatCardHoverState {
-  value: number | null;
-  label: string | null;
-  trend: number | null;
-}
-
-export const statCardValueClassName =
-  "text-3xl font-semibold leading-none tracking-tight";
-
-export const statCardLabelClassName = "mt-0 text-xs";
