+++ apps/web/public/r/stat-card-area-01-example.json
@@ -22,7 +22,7 @@
     },
     {
       "path": "registry/examples/stat-card-area-01-index.ts",
-      "content": "/** biome-ignore-all lint/performance/noBarrelFile: v0 registry example barrel for shadcn install */\n\nexport { LinearGradient } from \"@visx/gradient\";\nexport { Area } from \"./area\";\nexport { AreaChart } from \"./area-chart\";\nexport { ChartStatFlow } from \"./chart-stat-flow\";\n",
+      "content": "/** biome-ignore-all lint/performance/noBarrelFile: v0 registry example barrel for shadcn install */\nexport { AreaChart } from \"./area-chart\";\nexport { Area } from \"./area\";\nexport { ChartStatFlow } from \"./chart-stat-flow\";\nexport { LinearGradient } from \"@visx/gradient\";\n",
       "type": "registry:lib",
       "target": "components/charts/index.ts"
     }
