+++ apps/web/public/r/stat-card-choropleth-01-example.json
@@ -22,7 +22,7 @@
     },
     {
       "path": "registry/examples/stat-card-choropleth-01-index.ts",
-      "content": "/** biome-ignore-all lint/performance/noBarrelFile: v0 registry example barrel for shadcn install */\n\nexport { ChartStatFlow } from \"./chart-stat-flow\";\nexport {\n  ChoroplethChart,\n  ChoroplethFeatureComponent,\n  ChoroplethTooltip,\n} from \"./choropleth\";\n",
+      "content": "/** biome-ignore-all lint/performance/noBarrelFile: v0 registry example barrel for shadcn install */\nexport { ChoroplethChart, ChoroplethFeatureComponent, ChoroplethTooltip } from \"./choropleth\";\nexport { ChartStatFlow } from \"./chart-stat-flow\";\n",
       "type": "registry:lib",
       "target": "components/charts/index.ts"
     }
