+++ packages/ui/scripts/generate-block-registry.mjs
@@ -87,7 +87,6 @@ const BLOCKS = {
       "components/stat-card-area.tsx",
       "components/stat-card-chart.tsx",
       "components/stat-card-hover-bridge.tsx",
-      "components/stat-card-stat-styles.ts",
       "components/trend-badge.tsx",
       "data/revenue-series.ts",
     ],
@@ -112,7 +111,6 @@ const BLOCKS = {
       "components/stat-card-line.tsx",
       "components/stat-card-chart.tsx",
       "components/stat-card-hover-bridge.tsx",
-      "components/stat-card-stat-styles.ts",
       "components/trend-badge.tsx",
       "data/sessions-series.ts",
     ],
@@ -138,7 +136,6 @@ const BLOCKS = {
       "components/stat-card-choropleth.tsx",
       "components/stat-card-choropleth-hover-bridge.tsx",
       "components/stat-card-chart.tsx",
-      "components/stat-card-stat-styles.ts",
       "components/trend-badge.tsx",
       "data/visitors.ts",
       "lib/use-world-data.tsx",
